Moved ChangeLogs to a specific folder
