+++ src/ChangeLogs/CHANGELOG_AFFLICTION_WARLOCK.js
