+++ src/ChangeLogs/CHANGELOG_ENHANCEMENT_SHAMAN.js
