+++ src/ChangeLogs/CHANGELOG_HOLY_PALADIN.js
