+++ src/Main/Changelog.js
@@ -1,19 +1,19 @@
 import React from 'react';
 
-import CHANGELOG_CORE from '../CHANGELOG_CORE';
-import CHANGELOG_HOLY_PALADIN from '../CHANGELOG_HOLY_PALADIN';
-import CHANGELOG_DISCIPLINE_PRIEST from '../CHANGELOG_DISCIPLINE_PRIEST';
-import CHANGELOG_HOLY_PRIEST from '../CHANGELOG_HOLY_PRIEST';
-import CHANGELOG_RESTORATION_DRUID from '../CHANGELOG_RESTORATION_DRUID';
-import CHANGELOG_GUARDIAN_DRUID from '../CHANGELOG_GUARDIAN_DRUID';
-import CHANGELOG_MISTWEAVER_MONK from '../CHANGELOG_MISTWEAVER_MONK';
-import CHANGELOG_WINDWALKER_MONK from '../CHANGELOG_WINDWALKER_MONK';
-import CHANGELOG_RESTORATION_SHAMAN from '../CHANGELOG_RESTORATION_SHAMAN';
-import CHANGELOG_ELEMENTAL_SHAMAN from '../CHANGELOG_ELEMENTAL_SHAMAN';
-import CHANGELOG_ENHANCEMENT_SHAMAN from "../CHANGELOG_ENHANCEMENT_SHAMAN";
-import CHANGELOG_AFFLICTION_WARLOCK from "../CHANGELOG_AFFLICTION_WARLOCK";
-import CHANGELOG_BREWMASTER_MONK from "../CHANGELOG_BREWMASTER_MONK";
-import CHANGELOG_VENGEANCE_DEMON_HUNTER from "../CHANGELOG_VENGEANCE_DEMON_HUNTER";
+import CHANGELOG_CORE from '../ChangeLogs/CHANGELOG_CORE';
+import CHANGELOG_HOLY_PALADIN from '../ChangeLogs/CHANGELOG_HOLY_PALADIN';
+import CHANGELOG_DISCIPLINE_PRIEST from '../ChangeLogs/CHANGELOG_DISCIPLINE_PRIEST';
+import CHANGELOG_HOLY_PRIEST from '../ChangeLogs/CHANGELOG_HOLY_PRIEST';
+import CHANGELOG_RESTORATION_DRUID from '../ChangeLogs/CHANGELOG_RESTORATION_DRUID';
+import CHANGELOG_GUARDIAN_DRUID from '../ChangeLogs/CHANGELOG_GUARDIAN_DRUID';
+import CHANGELOG_MISTWEAVER_MONK from '../ChangeLogs/CHANGELOG_MISTWEAVER_MONK';
+import CHANGELOG_WINDWALKER_MONK from '../ChangeLogs/CHANGELOG_WINDWALKER_MONK';
+import CHANGELOG_RESTORATION_SHAMAN from '../ChangeLogs/CHANGELOG_RESTORATION_SHAMAN';
+import CHANGELOG_ELEMENTAL_SHAMAN from '../ChangeLogs/CHANGELOG_ELEMENTAL_SHAMAN';
+import CHANGELOG_ENHANCEMENT_SHAMAN from "../ChangeLogs/CHANGELOG_ENHANCEMENT_SHAMAN";
+import CHANGELOG_AFFLICTION_WARLOCK from "../ChangeLogs/CHANGELOG_AFFLICTION_WARLOCK";
+import CHANGELOG_BREWMASTER_MONK from "../ChangeLogs/CHANGELOG_BREWMASTER_MONK";
+import CHANGELOG_VENGEANCE_DEMON_HUNTER from "../ChangeLogs/CHANGELOG_VENGEANCE_DEMON_HUNTER";
 
 const changelogs = {
   core:CHANGELOG_CORE,
